Record sessions
Currently, I'm recording some anonymous information on each page visit.
But I'd like to start recording some more generic information that
should be true throughout one "session", such as which country the user
is from, and whether they're viewing the page on mobile or on
a computer.

Given the following timeline:

    |
    |> Initial page visit to `/`
    |  - Session recorded
    |    - Window dimensions
    |    - Approximate location (country/region)
    |    - Maybe more in future?
    |  - Visit recorded
    |    - Page visited (`/` in this case)
    |
    |> Visit new page `/articles`
    |  - Session NOT recorded
    |  - Visit recorded
    |

Some work is still needed. Ideally I'd like to only record this
information if `do not track` is disabled, and there is a lot of room
for improvement in the implementation of the lambda functions.