Add blobs to landing page and rework colours
After viewing the colour palette on my site with fresh eyes, I released
that things could use some work. More iteration is needed, but this is
a great start.

I also found the landing page to be quite bland, so I've experimented
with more colour and a few abstract blobs.

colour-rework